Add CRUD in Product

+++ app/Http/Controllers/CategoryController.php
@@ -21,7 +21,7 @@ class CategoryController extends Controller
         ]);
 
         $request->request->add(['slug' => $request->name]);
-        Category::created($request->except('_token'));
+        Category::create($request->except('_token'));
         return redirect(route('category.index'))->with(['succes' => 'Kategori Baru Ditambahkan!']);
     }
 
@@ -49,8 +49,8 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
-        $category = Category::withCount(['child'])->find($id);
-        if ($category->child_count == 0) {
+        $category = Category::withCount(['child', 'product'])->find($id);
+        if ($category->child_count == 0 && $category->product_count == 0) {
             $category->delete();
             return redirect(route('category.index'))->with(['success' => 'Kategori Dihapus!']);
         }
